Fix mock error
